add karma report in browser, improve php tests with smaller fixtures loaded (so it improves performances)
