+++ package.json
@@ -30,7 +30,7 @@
         "test-cafe": "./node_modules/.bin/testcafe $npm_package_config_test_browser ./assets/tests  --reporter spec,xunit:var/report/testcafe.xml",
         "test-cafe:windows": "./node_modules/.bin/testcafe %npm_package_config_test_browser% ./assets/tests  --reporter  spec,xunit:var/report/testcafe.xml",
         "test-karma": "./node_modules/.bin/karma start assets/tests/units/karma.conf.js --single-run",
-        "test-watch-karma": "./node_modules/.bin/karma start assets/tests/units/karma.conf.js",
+        "test-karma-watch": "./node_modules/.bin/karma start assets/tests/units/karma.conf.js",
         "test-php": "./vendor/bin/phpunit",
         "test-php:windows": ".\\vendor\\bin\\phpunit.bat",
         "test": "npm run test-php && npm run test-karma && npm run test-cafe",
@@ -46,6 +46,7 @@
         "axios": "^0.18.0",
         "graphql": "^0.13.1",
         "graphql-tag": "^2.8.0",
+        "karma-jasmine-html-reporter": "^0.2.2",
         "prop-types": "^15.6.0",
         "quasar-extras": "^1.0.2",
         "quasar-framework": "^0.14.9",
